feat(Dropdown): add a11y prop to tab moves focus (#1600)
This adds a new a11y prop to Dropdrop, defaults to true.
When the a11y prop is explicitly set to false:
This changes the way tab works within the Dropdown.
tab/shift-tab will move focus to the next logic item:
tab will move down until the end. Then, at the end of the menu it will focus the next focusable item on the page and close the menu.
shift-tab will move up until the start. Thenm, at the beginning of the menu it will focus the toggle and close the menu.

Fixes #1441